Renames API folders and minor changes to autocomplete component
